Implemented Interpreter
The interpreter class was implemented. We also included the entire TUIO
library and all files reuiqred by it (note this is not our work!).
Changed the TUIOProcessing class so as to incorporate the Interpreter
class without having to make the Interpreter extend a Papplet.
We also added a .gitignore to ignore all the .class files.
Will remove TUIO.jar in next commit since it is now redundant.