Added scrollTrigger, slotName and verbose level
Added a scrollTrigger which controls the number of degrees of rotation
of the scroll command block that will trigger a slot change.
Added a slotName, which is the default name of the slot files (with the
appropriate slot number appended).
Added a verbose level. 0 is the user default, 3 is the current maximum.
Higher means more verbose.