Further implementation of MysticalModelCenterCalculator
The MMCC has gotten so far as to move d to the origin, align model and
world d and make a new vector space with d, offset and their cross
product a basis.
We next need to find the model center in d-space so that we can
eliminate half of the possibilities easily.
We then need to do the part that converts orientations back to world
space.